+++ src/mobx.ts
@@ -80,7 +80,7 @@ export { isComputed, isComputedProp } from "./api/iscomputed"
 export { extendObservable, extendShallowObservable } from "./api/extendobservable"
 export { observe } from "./api/observe"
 export { intercept } from "./api/intercept"
-export { autorun, autorunAsync, when, reaction, IReactionOptions } from "./api/autorun"
+export { autorun, when, reaction, IReactionOptions } from "./api/autorun"
 export { action, isAction, runInAction, IActionFactory } from "./api/action"
 
 export { expr } from "./api/expr"